[gardening] Split ADTTests.cpp into separate test files.
