+++ unittests/Basic/CMakeLists.txt
@@ -6,7 +6,9 @@ handle_gyb_sources(
     ${SWIFT_HOST_VARIANT_ARCH})
 
 add_swift_unittest(SwiftBasicTests
-  ADTTests.cpp
+  ValueEnumeratorTest.cpp
+  OptionSetTest.cpp
+  RangeTest.cpp
   BlotMapVectorTest.cpp
   ClusteredBitVectorTest.cpp
   Demangle.cpp
